+++ packages/react-core/src/components/atoms/Input/PressableInputContainer/styled.ts
@@ -1,0 +1,5 @@
+import styled from '@emotion/native';
+
+export const StyledPressableSurface = styled.Pressable`
+  width: 100%;
+`;