+++ packages/react-core/src/components/molecules/DatePicker/styled.ts
@@ -1,10 +1,11 @@
 import styled, { css } from '@emotion/native';
+import { FC } from 'react';
 import { Platform } from 'react-native';
 import { StyleProps } from '../../../types/defaults';
-import {
-  PressableSurface,
-  PressableSurfaceProps,
-} from '../../atoms/PressableSurface';
+import { Icon } from '../../atoms/Icon';
+import { disabledInputStyles, InputContainerProps } from '../../atoms/Input';
+import { PressableSurface } from '../../atoms/PressableSurface';
+import { TextProps } from '../../atoms/Text';
 
 const isWeb = Platform.OS === 'web';
 
@@ -23,8 +24,11 @@ export const Backdrop = styled(PressableSurface)`
       `}
 `;
 
-export const ModalContent = styled.View<Partial<StyleProps>>`
+export const ModalContent = styled.View<
+  { offset: number } & Partial<StyleProps>
+>`
   background-color: transparent;
+  padding-bottom: ${({ offset }) => `${offset}px`};
   ${({ theme: { borderRadius } }) => css`
     ${isWeb
       ? `
@@ -36,4 +40,18 @@ export const ModalContent = styled.View<Partial<StyleProps>>`
     `}
   `}
   overflow: hidden;
+`;
+
+export const getStyledTextComponent = (component: FC<TextProps>) => {
+  return styled(component)(
+    (props: Partial<InputContainerProps> & Partial<StyleProps>) => css`
+      line-height: ${props.theme?.typography.h5.lineHeight};
+      ${disabledInputStyles(props)}
+    `
+  );
+};
+
+export const CalendarIcon = styled(Icon)<Partial<StyleProps>>`
+  padding: ${({ theme }) => theme.spacing.centi};
+  color: ${({ theme }) => theme.color.secondary.medium};
 `;